PG - Edit Data Changes (#117)
Fixed the feedback except the using psycopg2 for generating the queries. We use the mogrify for psycopg which expects the query template to be in certain format. As of now we manually handle escaping any column names of table names. But I will make this change to use SQL from the psycopg2