Integrate PG-CLI completion code into tools service (#121)
This commit integrates the bulk of the PG-CLI code for auto-completion into our codebase. Putting this out as its own PR for clarity, since the separate work to integrate into the language service will be much smaller. The code has been changed to:

- Avoid supporting CLI-only special syntax
- Make unit tests support the unittest framework rather than pytest
- Integrate with PGSMO in the CompletionRefresher class

**Additional changes**

Added a ThirdPartyNotices.txt with our 2 codebases we're using directly in the code. We'll also want to update this to a) be included in the bundle so it's installed on disk with the exe, and b) add in any module dependencies used at runtime (not test dependencies)
Notes

**Notes**
The code under pgsqltoolsservice\language\completion has been minimally changed from PG-CLI, and should be updatable in the future for changes. Similarly, unit tests under tests\language\completion have been changed from pytest to nosetest + parameterized. 

Long-term will look at whether we can go to a direct pg-cli dependency but this requires investigation re: only building in the modules we need, since it has a lot of extra dependencies, and filtering out / avoiding paths where the special CLI handling is processed. These are candidates for submitting back as config options to their codebase once we're public, since at that point they might be willing to take this and let us factor this cleanly out.