Fix query editor connect/disconnect
When connection is closed by server, the qurery execution fails but we were trying to set autocommit to true on a closed connection which throws an error.
Now check the status of the connection. This will ensure that the status is set to executed and user can connect/disconnect properly.