First steps toward a new endpoint for copilot
The ADS copilot will need a SQL description of the schema in the form of
a series of CREATE TABLE statements. This begins the creation of a
endpoint -- tentatively named metadata/sql -- to provide that info.

This commit creates the endpoint to return a hardcoded string. My code
is cargo-culted from the surrounding patterns and feels unnecessarily
repetitive.

To test it manually, send this payload

    Content-Length: 96

    {
	    "jsonrpc": "2.0",
	    "id": 1,
	    "method": "metadata/sql",
	    "params": {
		    "owner_uri": "foo"
	    }
    }

to stdin of `python3 ossdbtoolsservice/ossdbtoolsservice_main.py`

The result is:

    Content-Length: 66

    {"id": 1, "jsonrpc": "2.0", "result": {"metadata": "Hello world"}}