Add ChatService (#501)
* Fix redefinition of _log_exception.

This was causing exceptions not to be logged, which made debugging more difficult.

* Annotate types; refactor deserialization to support Pydantic models

* Log errors that occur in writer

* Add async runner to service provider

* Allow deserializing pydantic models

* Add vscode params to arg parse

Fixes error starting server from vs code.

* Type annotation to reader

* Add logging to ObjectExplorerService

* ChatService implementation.

* Use released version of semantic kernel

Previously it was pointed to my branch, but wasn't actually using the postgres vector store, so we can use an official release.

* Replace requirements.txt with package install.

This solves issue with running in dev mode with vs code extension requiring copying ossdbtoolsservice_main.py out to the top level of the repo.

* (temp) test release

* Fix param in build.ps1

* Add missing [binary, pool] for psycopg

* Fix exception call

* Don't copy nonexistent file

* Use psycopg 3.2.4

* Use windows-2022 runner

* Refactor JSONRPCServer -> MessageServer

Main motivation here is to create separate code paths for pyinstaller if using a web server vs json rpc server. The pyinstaller targeting vscode was including ssl, which produced some issues on mac osx arm in loading openssl shared libraries. These were only required for the web server, and this refactor produces different code paths for web server vs rpc server context such that ssl and other web server dependencies will not be included in the pyinstaller.

* Avoid bundling unused pg_exe binaries

* Bundle lib-ssl in macosx arm

* Set macosx_deployment_target

* Pin numpy

This pins numpy to an older version that was built latest on macosx 11 in order to force support for older versions of mac. With 2.2.2, the mac 14 wheel is pulled because the runner is mac 14; this is causing incompatibility with older OS's.

* Remove branch trigger for dev-build

* Linting

* Fix tests