Add copilot queries to query history (#510)
* Update AsyncRunner usage; simplify message sending in servers and contexts.

This includes getting rid of RpcRequestContext/RpcNotificationContext, which just uses base class functionality. Web*Context overrides the base class behavior because a particular session is needed and the server implementation doesn't have that detail.

* Update services with new base class

* Update get_async_runner; cleanup code in chat_service.py

* Add notification for query history

* Tweaks to system prompt

* Update/fix unit tests with server/context refactors.

Also add type annotations various places to trigger editor type error help. Formatting changes.

* Test send_request; Create tests_v2 to hold new pytests.

* Remove unused requirements.txt

* Add PR action

* Update links for main rename

* Move off of ubuntu-20.04

https://github.com/actions/runner-images/issues/11101

* Update PR build

* Fix tests

* Test host runners in PR CI

* Add pytest to test-all

* Move test_v2 to tests_v2

* PR runs against ubuntu; nightly runs on more hosts

* Remove pytest.ini

This was applicable when pytest_asyncio was used; we are using the AsyncRunner to execute async code and so do not need pytest_asyncio

* Add linting to PR check

* Linting

* Add notes on linting and lint.ps1