Add mypy check, fix types (#512)
* Run ruff check --fix after ruff format

Otherwise ruff check can fail on things ruff format fixes.

* Add mypy to lint script.

Only runs over ossdbtoolsservice and tests_v2. Does not follow imports, which would end up checking other projects. This is to scope the linting to where new development will happen.

Skips checking ossdbtoolsservice.language.completion. This module was particularly tricky to fix with typing. Could be tackled in future, but for now allow this to continue to be untyped.

* Pass mypy type checking.

* PostgreSQLConnection -> ServerConnection