Fix resultSet optionality for queries with errors
Prior to changes applying stricter type checking, batch query execution
was allowed to return None for result set summaries to the client. That
case was accidently prevented and threw a ValueError, which prevented
the actual DB error from being included in the response to the client. A
test was added to ensure this behavior is restored and remains.