Fix script request contract for deserialization
The ObjectMetadata attribute type for scripting object originally had a
hacky workaround because the client/server contracts were misaligned for
making a script request. When that was removed for typing fixes, the
parameter was no longer getting deserialized into the correct type.

Rather than fix the workaround, the contracts were sorted out, with
correct typing, and now the common deserialization code works correctly.
This also requires a change on the vscode client side to align.