Correct query exception details and fixes for "Script As" requests (#514)
* Fix resultSet optionality for queries with errors

Prior to changes applying stricter type checking, batch query execution
was allowed to return None for result set summaries to the client. That
case was accidently prevented and threw a ValueError, which prevented
the actual DB error from being included in the response to the client. A
test was added to ensure this behavior is restored and remains.

* Fix attach-to-process debugging

The extra configuration to debugpy was preventing attaching to the
process for breakpoint debugging.

* Fix script request contract for deserialization

The ObjectMetadata attribute type for scripting object originally had a
hacky workaround because the client/server contracts were misaligned for
making a script request. When that was removed for typing fixes, the
parameter was no longer getting deserialized into the correct type.

Rather than fix the workaround, the contracts were sorted out, with
correct typing, and now the common deserialization code works correctly.
This also requires a change on the vscode client side to align.