Consolidate sql.SQL wrap to as_sql
Pylance shows a type error for passing a string into cursor.execute and other methods, which are typed expecting psycopg.sql.SQL or variants. Also, sql.SQL is typed as a Literal[string], and so type errors would occur for non-literal strings. This change pushes the # type:ignore to a single place and uses the as_sql method to convert from str -> sql.SQL.