OE session mapping accounts for default connection values (#520)
A recent change to the client behavior meant that the object explorer
would send a standard ownerUri based on connection details. Previously,
the inconsistency meant that the map of connections tracked by the
connection manager actually had two different connections stored.

Now that a single URI was used, the tools service should have been able
to just use the single connection. However, the implementation caused a
bug and consistent race condition. When the connection manager got the
request the for second connection request, it would now correctly see
that it already had a connection for these details (via the URI). It
then checked to see if the connection details had changed. Because the
incoming details didn't have the default values (like port, dbname)
applied to it, but did have them on the details stored from the
connection that was actually opened, it incorrectly believed it to be a
"new" connection for the same server. It then closed the connection to
re-create it later.

Meanwhile, the client has likely requested nother operation involving
the connection, which was possibly closed, resulting in a failure. By
extending the equality check to account for default values, the tools
service should no longer close connections when the details haven't
actually changed.