Fixs bugs in EXPLAIN ANALYZE and ServerConnection parameters (#521)
* Fix typo in EXPLAIN ANALYZE template

The "show actual plan" option has been set to the incorrect syntax
`ANALYZE EXPLAIN` since it was committed about 8 years ago. This has
been swapped with the correct syntax `EXPLAIN ANALYZE` and a test has
been added to confirm.

* Map tool service options to PG options

There was a bug in translating connection properties from some internal
names to the ones psycopg expect. This resulting in properties like
`connect_timeout` from not being applied to the connection.