OE still uses "get_connection"
This creates an "orphaned" connection that is disconnected from a PooledConnection context. Orphaned connections can still be returned to their pool, though this will not be tracked by the ConnectionManager.

OE still holding a long-lived connection was due to the complexity of changes to refactor OE to use PooledConnection, specifically in pgsmo and smo. This means for every ConnectionDetail connected to, there will be 2 connections min - one from the pool, and one for OE. This is still an improvement. Will leave a refactor of OE to PooledConnection for future work, if necessary.