Remove message content from logs
The json message content of the LSP contain values directly from a user
database. These messages shouldn't be logged or persisted anywhere, so
the entire content of the message has been removed.

Also updates the default value for client_encoding to instruct clients
to prefer utf8 if not user value is specified.