Remove formatting for scripting service (#534)
* Remove formatting for scripting service

The scripting service recently was changed to include sqlparse
formatting prior to returning the generated sql. However, this
formatting doesn't produce well formatted results in many cases.
Instead, the sql templates that were producing the original poorly
formatted results were changed to be formatted nicely, as all other
scripting output is.

The formatting is still an issue in generally, and that is being tracked.

* Add improved docstring to VSCodeChatCompletion (#532)

* Patch the smoke tests

---------

Co-authored-by: Rob Emanuele <rdemanuele@gmail.com>