Refactor connection management (#536)
* Remove db exception abstraction

* Remove ConnectionManager, was not useful wrapper

* Inline server_type

* Removed unused connection_options prop

* Refactor connetion service to use connection pools.

Remove the 'driver' submodule, organize all core connection logic into connections.core. Introduce a ConnectionManager that is responsible for managing connection pools, what owner_uris are associated with them, connections in active transactions, and more.

* Move to using PooledConnection in various services

* Use PooledConnection in language service

* OE still uses "get_connection"

This creates an "orphaned" connection that is disconnected from a PooledConnection context. Orphaned connections can still be returned to their pool, though this will not be tracked by the ConnectionManager.

OE still holding a long-lived connection was due to the complexity of changes to refactor OE to use PooledConnection, specifically in pgsmo and smo. This means for every ConnectionDetail connected to, there will be 2 connections min - one from the pool, and one for OE. This is still an improvement. Will leave a refactor of OE to PooledConnection for future work, if necessary.

* Move query service to PooledConnection

* Don't rollback user queries. Force them to roll back, allowing user debugging.

* Make note of unused messages and parameters in vscode

* Cleanup - remove unused code, formatting

* Update main to use service_provider as context manager

* Update tests

* Update tests_v2

* Move query editor to use a long lived connection.

This is due to considerations on the query editor being able to set connection-level properties like GUC or temporary tables. These shouldn't be shared amongst query editors, and so the connection for a query editor window is really tied to that editor only.

* Updates on progress moving query to dedicated connection

Not quite working yet

* Update upstream changes with connection pool changes.

* Use sorted params to make connection details hash

* Don't use out-of-band transaction error tracking

* Fix ResultMessage batch_id type

* Set min_size=0, max_idle and max_lifetime for ConnectionPool

* Fix typo in private attribute

* Make _details_to_owner_uri map to a set of uris

* Add connection/transfer request.

Also remove unused switch database request

* Format

* Lint

* Test fixes.

* Track long lived connections by owner_uri, not also by type/name

This fixes an implementation that held long-lived connections seperate for owner_uris based on the incoming connection_name or type. This was breaking the transfer logic. There should be a one-to-one between a long-lived connection and an owner_uri.

* Fix azure token handling in connection class

This was correct beforehand, but didn't use the static method that was extracted so that testing would use same logic as normal code flow. Renamed to make more clear it may handle the azure token, but not if there's no token

* Refactor configuration; drop sql config

This change extracts Configuration into it's own file, and makes a note of how to add additional configuration that gets synced to vscode. Adds max_connections to configuration.

* ConnectionManager users configuraiton for max pool size.

* Fix MessageServerClientWrapper.send_client_notification

* Formatting

* Fix test