Fix chat history manager issue with message hashing.
Previously, the user message hash was used to identify a spot in a conversation history that the tool call history should be inserted. This presented an issue whe messages from the user were the same - particularly, "yes" is a common message because the model often asks "would you like me to proceed?". This would cause an error when the chat history manager attempted to insert a tool call history multiple times.

Instead, we use the index of the user message in the chat history as the identifier. This should remain consistent across CompletionRequest calls as the history is append-only for a single session.