Avoid prepared statements in pooled connections.
Errors were popping up around psycopg-generated prepared statements not existing. This is likely due to the sharing of connections via the connection pool in different scenarios. This change avoid psycopg from automatically triggering a prepared statement when a statement occurs 3 times. This may result in some performance hit for queries that happen over and over, but avoid correct queries returning errors. We'll need to manually prepare statements when using pooled connections if we want that speedup.