Chat fixes (#537)
* Fix chat history manager issue with message hashing.

Previously, the user message hash was used to identify a spot in a conversation history that the tool call history should be inserted. This presented an issue whe messages from the user were the same - particularly, "yes" is a common message because the model often asks "would you like me to proceed?". This would cause an error when the chat history manager attempted to insert a tool call history multiple times.

Instead, we use the index of the user message in the chat history as the identifier. This should remain consistent across CompletionRequest calls as the history is append-only for a single session.

* Avoid prepared statements in pooled connections.

Errors were popping up around psycopg-generated prepared statements not existing. This is likely due to the sharing of connections via the connection pool in different scenarios. This change avoid psycopg from automatically triggering a prepared statement when a statement occurs 3 times. This may result in some performance hit for queries that happen over and over, but avoid correct queries returning errors. We'll need to manually prepare statements when using pooled connections if we want that speedup.